fix up esri-json importer
cleaned up the importer to function with points polylines and polygons
